snmp.version: move unittests to tests/
